refactor structure into packages
