+++ src/main/java/chatbot/interfaces/AbstractTask.java
@@ -1,4 +1,4 @@
-package main.java.interfaces;
+package chatbot.interfaces;
 
 public interface AbstractTask {
     void setDone(boolean status);
