+++ src/main/java/chatbot/interfaces/TaskStorage.java
@@ -1,4 +1,4 @@
-package main.java.interfaces;
+package chatbot.interfaces;
 
 public interface TaskStorage<T> {
     TaskStorageResult<T> getTasks();
