+++ src/main/java/chatbot/interfaces/TaskStorageResult.java
@@ -1,4 +1,4 @@
-package main.java.interfaces;
+package chatbot.interfaces;
 
 public interface TaskStorageResult<T> {
     void showResult(MessageView<T> messageView);
